requirements.txt
openai
google.generativeai
flask
supabase
bcrypt
python-dotenv
azure-ai-inference